requirements.txt
django
django-ldapdb
bleach
python-creole
requests
oauthlib
urllib3
certifi
idna
chardet
django-auth-ldap
django-oauth-toolkit
django-cors-headers
django-widget-tweaks
etherpad_lite
fuzzywuzzy
nameparser
networkx
psycopg2
pygpgme
pygraphviz
pytz
pyPDF2
raven
pyldap
python-magic
pydiscourse
django-email-log
pillow
